requirements.txt
streamlit>=1.49
pypdf>=4.0.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-chroma>=0.2.0
langchain-google-genai>=2.1.0
langchain-text-splitters>=0.3.0
chromadb>=1.0.0
pyarrow>=15.0.0
python-dotenv>=1.1.1
pypdf
huggingface
langchain-huggingface